Add drive train encoder
We only have three encoders right now, and they're more important on
both sides of our lift than on the drive train for now, so there's only
one encoder on it.

Remove sensor::init. It would be better to inline the function in
init.cpp depending on whether it should be called during initializeIO()
or initialize().

Add an AnalogPort enum in case we start using them.